requirements.txt
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==1.11.6
django-cors-headers==3.1.1
django-crispy-forms==1.7.2
django-filter==2.2.0
django-filters==0.2.1
django-formtools==2.1
django-guardian==2.1.0
django-redis==4.10.0
django-reversion==3.0.4
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
future==0.17.1
httplib2==0.14.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
mysqlclient==1.4.4
oauthlib==3.1.0
Pillow==6.2.0
pycryptodome==3.9.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.2
raven==6.10.0
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
uritemplate==3.0.0
urllib3==1.25.6
XlsxWriter==1.2.1
xlwt==1.3.0
